+++ src/app/place/components/index.ts
@@ -7,3 +7,4 @@ export { default as Info } from './Info';
 export { default as ReviewSheet } from './ReviewSheet';
 export { default as TimeLogSheet } from './TimeLogSheet';
 export { default as ShareSheet } from './ShareSheet';
+export { default as PlaceTopNavigationBar } from './PlaceTopNavigationBar';
